bacnet/epics: Skip the "device" object in the object-list
The "device" object is always the first object in the object-list, so we
can safely skip that first entry. Reading the object-list property is
part of reading the device object, so keeping it on the list is just
going to read the same object twice.

The consequence is also that devices that don't have any data points in
them (like routers) will return an empty object-list.